fixed the persistant empty todo entry at the top
Following the instructions strictly meant that the list started with one empty
element, which seems fine, and was necessary to pass the tests. But the element
stuck around as a blank spot at the top of the list as more TODOs were appended,
which was ugly. This commit makes the first TODO overwrite the initial blank
one; subsequent TODOs are appended to the list's bottom as normal.